+++ apps/cowswap-frontend/src/modules/limitOrders/services/tradeFlow/index.ts
@@ -1,4 +1,4 @@
-import { reportPermitWithDefaultSigner } from '@cowprotocol/common-utils'
+import { captureError, ERROR_TYPES, normalizeError, reportPermitWithDefaultSigner } from '@cowprotocol/common-utils'
 import { SigningScheme } from '@cowprotocol/cow-sdk'
 import { Percent } from '@cowprotocol/currency'
 import { isSupportedPermitInfo } from '@cowprotocol/permit-utils'
@@ -183,12 +183,13 @@ export async function tradeFlow(
     analytics.sign(swapFlowAnalyticsContext)
 
     return orderId
-    // TODO: Replace any with proper type definitions
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  } catch (error: any) {
+  } catch (err: unknown) {
+    const error = normalizeError(err)
+
     logTradeFlow('LIMIT ORDER FLOW', 'STEP 9: ERROR: ', error)
     const swapErrorMessage = getSwapErrorMessage(error)
 
+    captureError(error, ERROR_TYPES.ON_SWAP, { swapErrorMessage })
     analytics.error(error, swapErrorMessage, swapFlowAnalyticsContext)
 
     throw error