+++ apps/cowswap-frontend/src/modules/tradeFormValidation/hooks/useTradeFormValidationContext.ts
@@ -13,8 +13,9 @@ import { useHasHookBridgeProvidersEnabled } from 'entities/bridgeProvider'
 import { useCurrentAccountProxy } from 'modules/accountProxy/hooks/useCurrentAccountProxy'
 import { useTokensBalancesCombined } from 'modules/combinedBalances'
 import { useApproveState, useGetAmountToSignApprove, useIsApprovalOrPermitRequired } from 'modules/erc20Approve'
+import { useInjectedWidgetParams } from 'modules/injectedWidget'
 import { RwaTokenStatus, useRwaTokenStatus } from 'modules/rwa'
-import { TradeType, useDerivedTradeState, useIsWrapOrUnwrap } from 'modules/trade'
+import { TradeType, useDerivedTradeState, useIsWrapOrUnwrap, useTradePriceImpact } from 'modules/trade'
 import { TradeQuoteState, useTradeQuote } from 'modules/tradeQuote'
 
 import { QuoteApiError, QuoteApiErrorCodes } from 'api/cowProtocol/errors/QuoteError'
@@ -31,6 +32,8 @@ export function useTradeFormValidationContext(): TradeFormValidationCommonContex
   const { account } = useWalletInfo()
   const derivedTradeState = useDerivedTradeState()
   const tradeQuote = useTradeQuote()
+  const injectedWidgetParams = useInjectedWidgetParams()
+  const tradePriceImpact = useTradePriceImpact()
   const isProviderNetworkUnsupported = useIsProviderNetworkUnsupported()
   const isProviderNetworkDeprecated = useIsProviderNetworkDeprecated()
   const isOnline = useIsOnline()
@@ -99,6 +102,8 @@ export function useTradeFormValidationContext(): TradeFormValidationCommonContex
       isRestrictedForCountry,
       isBalancesLoading: !hasFirstLoad || isBalancesLoading,
       balancesError,
+      injectedWidgetParams,
+      tradePriceImpact,
       isInputCurrencyXstock,
       isOutputCurrencyXstock,
     }
@@ -124,11 +129,13 @@ export function useTradeFormValidationContext(): TradeFormValidationCommonContex
     isWrapUnwrap,
     isProxySetupValid,
     isInputCurrencyXstock,
+    isOutputCurrencyXstock,
     recipientEnsAddress,
     toBeImported,
     tradeQuote,
     balancesError,
-    isOutputCurrencyXstock,
+    injectedWidgetParams,
+    tradePriceImpact,
   ])
 }